+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-## [2.0.5-dev.1] - 2019-03-10
+## [2.0.5-dev.2] - 2019-04-17
+
+- fix es6 with dynamic import to not show as missing dependencies
+
+## [2.0.5-dev.1] - 2019-04-16
 
 - improve performance by lazy load all external packages and internal files